Add files via upload
This kaggle project involves classifying the tweets as disaster(1) or non disaster(0).A total of 10,000 tweets has to be identified.Basic Exploratory data analysis is done to find the text and token length in each of these tweets, printing word clouds etc. After data preprocessiong,Data is trained on different models to find the best F1 score and hyperparameter tuning is done.The model is trained on the best F1 score model and submitted to find the accuracy score.